Refactor partial order generation code.
Use a simpler implementation that follows the symmetry breaking paper
exactly, except that instead of using the largest eclass, you use the
first one. Since the patterns are sorted by degree and neighbour degree,
the earlier eclasses will contain vertices that comprise the pattern
core. If conditions are first placed on core vertices, you won't have
conditions spanning sibling groups.

Additionally, we need to handle the case that an order group straddles a
core vertex. This only happens in patterns with anti-edges. E.g.
[1-2][1-3][1-4](2~3)(3~4), where 2 < 3 < 4 and there is only one sibling
group. This is a problem because 2's lower bound is 4's upper bound, but
they're in the same order group. This leads to no matches being found. To
deal with this, we need to split up the order group that normally
occurs, e.g. {2, 4}, into two, and tell the matcher not to find an upper
bound for the sibling group when computing its candidate set.