Improve DataGraph handling by match_* functions
Previously, to support both SmallGraph and directory name constructors
to DataGraph, the match() function passed its DataGraphT parameter to
each of the match_* functions, which would read the data graph in
separately.

Now, match() reads the DataGraph in first and stores it in the Context,
which gets accessed by the match_* functions. DataGraphT can now also
just be DataGraph, in which case it gets stored in the Context without
having to read anything. More DRY, and no more redundant data graph
reads.