Update sample FSM app
The original FSM app was written for edge-induced only and had redundant
data graph reads. Now users can choose vertex- or edge-induced as a
command line argument and it only reads the data graph in once  Now
users can choose vertex- or edge-induced as a command line argument and
it only reads the data graph in once.

Default is still edge-induced, so existing scripts that call the app
shouldn't be affected. But you can now do like so:

```
$ bin/fsm data/graph 3 100 v   # vertex-induced
$ bin/fsm data/graph 3 100 e   # edge-induced
$ bin/fsm data/graph 3 100 v 8 # vertex-induced, 8 threads
$ bin/fsm data/graph 3 100 8 v # vertex-induced, 8 threads
```