Drastically simplify to_string()
Previously, to_string() was overloaded to print the pattern given a
labeling. This seemed unnecessary, so the overloads were combined into
one method. Additionally, the labeled code did some complicated (and
buggy) sorting to try and give canonical output, because of how some
tests used to be implemented. This is done away with, since it caused
output to be misleading. Now the string output matches the actual
pattern vertices.