Re-implement early termination using jthread
Previously, Peregrine used pthread_cancel() to implement early
termination. This was a bit of an unstructured disaster and required
care throughout the remaining codebase. C++20 has jthreads, which can be
requested to stop by any thread, and zero-cost exceptions, so this
commit uses jthreads along with a barebones custom exception to achieve
the same results in a cleaner manner.

Along with this change, the thread local values are no longer stored in
a thread-local space, but managed on the heap by the aggregator. This
means that on worker exit we don't lose aggregation values, so there is
no need to synchronize worker exits with aggregator access.

Finally, this does mean that g++-9 is no longer supported, as jthreads
were implemented as of libstdc++ 10.1. See more here:
https://gcc.gnu.org/onlinedocs/libstdc++/manual/status.html#status.iso.2020