Write vertex at block boundary before flushing
Previously, when using the CSV format in the OutputManager, if a vertex
aligned with the output block size, it wouldn't be written to disk
since the output buffer size was not updated before flushing.

This commit updates the buffer size before flushing, so the last vertex
is never skipped in the output.